add: policy

+++ app/Http/Controllers/FakultasController.php
@@ -32,6 +32,10 @@ class FakultasController extends Controller
      */
     public function store(Request $request)
     {
+        // cek apakah user memiliki izin untuk membuat fakultas
+        if ($request->user()->cannot('create', Fakultas::class)) {
+            abort(403);   
+        }
         // validasi input
         $input = $request-> validate([
             'nama' => 'required|unique:fakultas',
@@ -72,6 +76,10 @@ class FakultasController extends Controller
     public function update(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
+        // cek apakah user memiliki izin untuk mengupdate fakultas
+        if ($request->user()->cannot('update', $fakultas)) {
+            abort(403);   
+        }
         //validasi input
         $input = $request-> validate([
             'nama' => 'required',
@@ -87,10 +95,14 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         // dd($fakultas);
         $fakultas = Fakultas::findOrFail($fakultas);
+        // cek apakah user memiliki izin untuk delete fakultas
+        if ($request->user()->cannot('delete', $fakultas)) {
+            abort(403);   
+        }
         // dd($fakultas);
         $fakultas->delete();
         return redirect()->route('fakultas.index')->with('success', 'Fakultas a.n. '. $fakultas->nama.' berhasil dihapus.');
